// src/file-service/src/models/HealthControllerCheck503Response.ts
/**
 * Terminology
 * Terminology API
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HealthControllerCheck200ResponseInfoValue } from '../models/HealthControllerCheck200ResponseInfoValue';
import { HttpFile } from '../http/http';

export class HealthControllerCheck503Response {
    'status'?: string;
    'info'?: { [key: string]: HealthControllerCheck200ResponseInfoValue; } | null;
    'error'?: { [key: string]: HealthControllerCheck200ResponseInfoValue; } | null;
    'details'?: { [key: string]: HealthControllerCheck200ResponseInfoValue; };

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "status",
            "baseName": "status",
            "type": "string",
            "format": ""
        },
        {
            "name": "info",
            "baseName": "info",
            "type": "{ [key: string]: HealthControllerCheck200ResponseInfoValue; }",
            "format": ""
        },
        {
            "name": "error",
            "baseName": "error",
            "type": "{ [key: string]: HealthControllerCheck200ResponseInfoValue; }",
            "format": ""
        },
        {
            "name": "details",
            "baseName": "details",
            "type": "{ [key: string]: HealthControllerCheck200ResponseInfoValue; }",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return HealthControllerCheck503Response.attributeTypeMap;
    }

    public constructor() {
    }
}