// src/file-service/src/models/InternalCptControllerImport200Response.ts
/**
 * Terminology
 * Terminology API
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { InternalCptControllerImport200ResponseResult } from '../models/InternalCptControllerImport200ResponseResult';
import { HttpFile } from '../http/http';

export class InternalCptControllerImport200Response {
    'result'?: InternalCptControllerImport200ResponseResult;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "result",
            "baseName": "result",
            "type": "InternalCptControllerImport200ResponseResult",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return InternalCptControllerImport200Response.attributeTypeMap;
    }

    public constructor() {
    }
}