// src/file-service/src/models/LoincControllerSearch200Response.ts
/**
 * Terminology
 * Terminology API
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { LoincDto } from '../models/LoincDto';
import { HttpFile } from '../http/http';

export class LoincControllerSearch200Response {
    'result'?: Array<LoincDto>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "result",
            "baseName": "result",
            "type": "Array<LoincDto>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return LoincControllerSearch200Response.attributeTypeMap;
    }

    public constructor() {
    }
}